updating naming to automatic lazy rendering and finishing initial helper code
