+++ priority-elements/automatic-lazy-rendering/wp-rocket-automatic-lazy-rendering-parameters/README.md
@@ -1,20 +1,20 @@
-# WP Rocket | Lazy Render Content Parameters
+# WP Rocket | Automatic Lazy Rendering Parameters
 
 ⚠ Only use this helper with WP Rocket versions 3.17 and later.
 
 
-### Changes the following Lazy Render Content parameters:
-- Enable or Disable Lazy Render Content optimization.
-- Process additional HTML tags.
+### Changes the following Automatic Lazy Rendering parameters:
 - Change the threshold for lazy rendering content. 
 - Change the depth of elements for the Lazy Render Content optimization.
+- Change the maximun number of HTML tags to process on each page.
+- Customize the HTML tags eligible for Automatic Lazy Rendering.
 
 
 ### 📝Manual code edit required before use!
 Please find the comments (// EDIT HERE) showing where to make changes.
 
 ### Documentation:
-[soon]
+https://docs.wp-rocket.me/article/1835-automatic-lazy-rendering
 
 ### To be used with:
 * Any setup where the default Lazy Render Contents settings do not work optimally.
